requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
altair==4.2.2
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
async-timeout==4.0.3
asyncpg==0.30.0
attrs==25.1.0
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dataclasses-json==0.6.7
decorator==5.2.1
distro==1.9.0
docx2txt==0.8
EditorConfig==0.17.0
entrypoints==0.4
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.8
ffmpy==0.5.0
filelock==3.17.0
frozenlist==1.5.0
fsspec==2025.2.0
ghp-import==2.1.0
gitdb==4.0.12
GitPython==3.1.44
gradio==5.16.0
gradio_client==1.7.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.28.1
idna==3.10
ipython==8.34.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsbeautifier==1.15.4
jsonpatch==1.33
jsonpickle==4.0.2
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.18
langchain-community==0.3.17
langchain-core==0.3.35
langchain-openai==0.3.5
langchain-text-splitters==0.3.6
langsmith==0.3.8
lxml==5.3.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mkdocs-mermaid2-plugin==1.2.1
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.26.0
networkx==3.3
numpy==1.26.4
openai==1.63.0
orjson==3.10.15
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pgvector==0.3.6
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.0
prompt_toolkit==3.0.50
propcache==0.2.1
protobuf==5.29.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.1
pymdown-extensions==10.14.3
pypdf==5.3.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.1
pyvis==0.3.2
PyYAML==6.0.2
pyyaml_env_tag==0.1
redis==5.0.3
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.23.1
ruff==0.9.6
safehttpx==0.1.6
scikit-learn==1.6.1
scipy==1.15.1
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
stack-data==0.6.3
starlette==0.45.3
streamlit==1.43.1
streamlit-mermaid==0.3.0
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.9.0
toml==0.10.2
tomlkit==0.13.2
toolz==1.0.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
watchdog==6.0.0
wcwidth==0.2.13
websockets==14.2
XlsxWriter==3.2.2
yarl==1.18.3
zstandard==0.23.0
